requirements.txt
PyGObject
pywebview
Flask